requirements.txt
fastapi
uvicorn
pydantic
openai
oracledb
pandas
python-multipart
python-dotenv
streamlit
numpy
scikit-learn
